Script is renamed and updated after clarifications
